# backend/swagger.yaml
openapi: 3.0.3
info:
  title: TETRA Project API
  version: 1.0.0
  description: |
    REST API documentation for TETRA Project.
    This API allows you to manage training modules, units, unit content (articles, videos, quizzes), questions, and answers.
servers:
  - url: http://localhost:8080
paths:
  /api/modules:
    get:
      summary: Get all modules
      description: |
        Returns a list of all training modules in the system.
        Each module contains information such as title, topic, description, points, cover image URL, and status.
      responses:
        '200':
          description: Successful response. Returns an array of Module objects.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Module'
    post:
      summary: Create a module
      description: |
        Creates a new training module.
        The request body must include all required fields for the module.
      requestBody:
        required: true
        description: Module object to be created.
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Module'
      responses:
        '201':
          description: Module successfully created. Returns the created Module object.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Module'
...
  /api/modules/{id}:
    delete:
      summary: Delete a training module by ID
      description: Deletes the module with the specified ID. **Admin access only.**
      tags:
        - TrainingModule
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          description: UUID of the module to delete
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Module deleted successfully
          content:
            application/json:
              schema:
                type: string
                example: Module deleted
        '401':
          description: Unauthorized - JWT is missing or invalid
        '403':
          description: Forbidden - User does not have ADMIN role
        '404':
          description: Module not found
...
  /api/users:
    post:
      summary: Create a new user
      description: |
        Creates a new user (ADMIN or LEARNER).
      security:
        - bearerAuth: []
      requestBody:
        required: true
        description: User creation payload.
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateUserRequest'
      responses:
        '201':
          description: User created successfully.
          content:
            application/json:
              schema:
                type: string
                example: "User created successfully."
        '400':
          description: Invalid input or error.
          content:
            application/json:
              schema:
                type: string
                example: "Error: Invalid role. Allowed roles: ADMIN, LEARNER"
        '500':
          description: Unexpected error.
          content:
            application/json:
              schema:
                type: string
                example: "Unexpected error: ..."

    get:
      summary: Get all users (ADMIN only)
      description: |
        Returns a list of all users. Requires ADMIN role.
      security:
        - bearerAuth: []
      responses:
        '200':
          description: List of users.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/UserSummary'
        '401':
          description: Missing or invalid Authorization header.
          content:
            application/json:
              schema:
                type: string
                example: "Missing or invalid Authorization header"
        '403':
          description: Access denied (not ADMIN).
          content:
            application/json:
              schema:
                type: string
                example: "Access denied"
  /api/users/{id}:
    put:
      summary: Update user by ID
      description: |
        Updates a user's name, email, and optionally password.
        If password is being changed, requires oldPassword for validation.
        Only ADMIN or the user themself can update.
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          description: The UUID of the user to update.
          schema:
            type: string
      requestBody:
        required: true
        description: User update payload.
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateUserRequest'
      responses:
        '200':
          description: User updated successfully.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserSummary'
        '400':
          description: Invalid input or error.
          content:
            application/json:
              schema:
                type: string
                example: "Error: Old password is incorrect"
        '401':
          description: Missing or invalid Authorization header.
          content:
            application/json:
              schema:
                type: string
                example: "Missing or invalid Authorization header"
        '403':
          description: Access denied.
          content:
            application/json:
              schema:
                type: string
                example: "Access denied"
        '404':
          description: User not found.
          content:
            application/json:
              schema:
                type: string
                example: "User not found"
        '500':
          description: Unexpected error.
          content:
            application/json:
              schema:
                type: string
                example: "Unexpected error: ..."

    get:
      summary: Get user by ID
      description: |
        Returns user information by user ID.
        Requires authentication (ADMIN or the user themself).
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          description: The UUID of the user to retrieve.
          schema:
            type: string
      responses:
        '200':
          description: User info.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserSummary'
        '404':
          description: User not found.
          content:
            application/json:
              schema:
                type: string
                example: "User not found"
        '401':
          description: Missing or invalid Authorization header.
          content:
            application/json:
              schema:
                type: string
                example: "Missing or invalid Authorization header"
        '403':
          description: Access denied.
          content:
            application/json:
              schema:
                type: string
                example: "Access denied"
...
components:
  schemas:
    Module:
      type: object
      description: |
        Represents a training module in the TETRA system.
        A module contains metadata and is associated with units and content blocks.
      properties:
        id:
          type: string
          description: Unique identifier (UUID) for the module.
          example: "123e4567-e89b-12d3-a456-426614174000"
        title:
          type: string
          description: The title of the module.
          example: "Introduction to Cybersecurity"
        topic:
          type: string
          description: The topic or category of the module.
          example: "Cybersecurity"
        description:
          type: string
          description: A detailed description of the module.
          example: "This module covers the basics of cybersecurity, including threats and best practices."
        points:
          type: integer
          description: The total number of points available for completing the module.
          example: 100
        coverUrl:
          type: string
          description: URL to the cover image for the module.
          example: "https://example.com/images/cybersecurity-cover.png"
        status:
          type: string
          description: The publication status of the module (e.g., 'draft', 'published').
          example: "published"
...
    CreateUserRequest:
      type: object
      description: Payload for creating a user.
      properties:
        name:
          type: string
          example: "John Doe"
        email:
          type: string
          example: "john@example.com"
        password:
          type: string
          example: "password123"
        role:
          type: string
          description: User role (ADMIN or LEARNER)
          example: "LEARNER"
    UserSummary:
      type: object
      description: Short info about a user.
      properties:
        id:
          type: string
          example: "123e4567-e89b-12d3-a456-426614174000"
        name:
          type: string
          example: "John Doe"
        email:
          type: string
          example: "john@example.com"
        role:
          type: string
          example: "LEARNER"
    UpdateUserRequest:
      type: object
      description: Payload for updating a user.
      properties:
        name:
          type: string
        email:
          type: string
        password:
          type: string
          description: New password (optional)
        oldPassword:
          type: string
          description: Required if password is being changed
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...